separate brain imaging from brain imaging samples types defs
