+++ server/routes/brainImaging.ts
@@ -1,7 +1,14 @@
 import fs from 'fs'
 import path from 'path'
 import serverconfig from '#src/serverconfig.js'
-import type { CategoricalTW, BrainImagingRequest, BrainImagingResponse, FilesByCategory, TermWrapper, RouteApi } from '#types'
+import type {
+	CategoricalTW,
+	BrainImagingRequest,
+	BrainImagingResponse,
+	FilesByCategory,
+	TermWrapper,
+	RouteApi
+} from '#types'
 import { brainImagingPayload } from '#types'
 import { spawn } from 'child_process'
 import { getData } from '../src/termdb.matrix.js'
@@ -13,6 +20,10 @@ export const api: RouteApi = {
 	endpoint: 'brainImaging',
 	methods: {
 		get: {
+			...brainImagingPayload,
+			init
+		},
+		post: {
 			...brainImagingPayload,
 			init
 		}
