+++ shared/types/src/index.ts
@@ -5,6 +5,7 @@ export * from './termsetting.ts'
 export * from './filter.ts'
 export * from './routes/routeApi.ts'
 export * from './routes/brainImaging.ts'
+export * from './routes/brainImagingSamples.ts'
 export * from './routes/burden.ts'
 export * from './routes/dataset.ts'
 export * from './routes/dsdata.ts'
